+++ test/OpenTelemetry.Instrumentation.Http.Tests/HttpClientTests.cs
@@ -387,7 +387,7 @@ public partial class HttpClientTests
 
             if (tc.RecordException.HasValue && tc.RecordException.Value)
             {
-                Assert.Single(activity.Events.Where(evt => evt.Name.Equals("exception")));
+                Assert.Single(activity.Events, evt => evt.Name.Equals("exception"));
                 Assert.True(enrichWithExceptionCalled);
             }
         }