+++ test/TestApp.AspNetCore/Program.cs
@@ -3,9 +3,7 @@
 
 using TestApp.AspNetCore;
 
-#pragma warning disable CA1515
 public class Program
-#pragma warning restore CA1515
 {
     public static void Main(string[] args)
     {