+++ test/TestApp.AspNetCore/TestActivityMiddleware.cs
@@ -3,9 +3,7 @@
 
 namespace TestApp.AspNetCore;
 
-#pragma warning disable CA1515
 public class TestActivityMiddleware
-#pragma warning restore CA1515
 {
     public virtual void PreProcess(HttpContext context)
     {