+++ src/OpenTelemetry.Instrumentation.AspNet/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 1.10.0-beta.1
+
+Released 2024-Dec-09
+
 * Updated registration extension code to retrieve environment variables through
   `IConfiguration`.
   ([#1976](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1976))